pyproject.toml
[tool.poetry]
name = "Ichigo"
version = "0.1.0"
description = ""
authors = ["Forbidden-A", "randomcmd"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hikari-lightbulb = "^2.0.1"
python-dotenv = "^0.19.2"
cattrs = "^1.8.0"
asyncpg = "^0.25.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
